+++ scripts/setup_dummy_signers.sh
