+++ src/commands/doge/bridge-init.ts
@@ -1,12 +1,9 @@
 import * as toml from '@iarna/toml'
-import { input, select } from '@inquirer/prompts'
+import { input } from '@inquirer/prompts'
 import { Command, Flags } from '@oclif/core'
 import chalk from 'chalk'
 import Docker from 'dockerode'
 import * as fs from 'node:fs'
-import * as path from 'node:path'
-import { loadDogeConfig } from '../../utils/doge-config.js'
-import type { DogeConfig } from '../../types/doge-config.js'
 import { getSetupDefaultsPath } from '../../config/constants.js'
 
 export class BridgeInitCommand extends Command {
@@ -14,6 +11,7 @@ export class BridgeInitCommand extends Command {
 
     static examples = [
         '$ scrollsdk doge:bridge-init',
+        '$ scrollsdk doge:bridge-init -s 123456',
         '$ scrollsdk doge:bridge-init --seed 123456',
     ]
 
@@ -27,7 +25,7 @@ export class BridgeInitCommand extends Command {
         const docker = new Docker();
         const image = `docker.io/dogeos69/generate-test-keys:v0.1.1-test`;
         try {
-            this.log(chalk.cyan('Pulling Docker Image...'))
+            this.log(chalk.cyan(`Pulling Docker Image: ${image}`))
             // Pull the image if it doesn't exist locally
             const pullStream = await docker.pull(image)
             await new Promise((resolve, reject) => {
@@ -72,7 +70,6 @@ export class BridgeInitCommand extends Command {
                     } else if (data.StatusCode !== 0) {
                         this.error(`Container exited with status code: ${data.StatusCode}`)
                     }
-
                     resolve(null)
                 })
             })
@@ -87,18 +84,24 @@ export class BridgeInitCommand extends Command {
     async run(): Promise<void> {
         const { flags } = await this.parse(BridgeInitCommand)
         let seed = flags.seed
-        if (!seed) {
-            seed = await input({
-                message: 'Enter the seed string',
-            })
-        }
 
+        // Read existing seed from setup_defaults.toml
         const setupDefaultsPath = getSetupDefaultsPath();
         if (!fs.existsSync(setupDefaultsPath)) {
             this.error('setup_defaults.toml not found, please run `scrollsdk doge:config` first')
             return
         }
         const existingConfigStr = fs.readFileSync(setupDefaultsPath, 'utf-8');
+        const existingConfig = toml.parse(existingConfigStr) as any;
+        const existingSeed = existingConfig.seed_string || '';
+
+        if (!seed) {
+            seed = await input({
+                message: 'Enter the seed string',
+                default: existingSeed,
+            })
+        }
+
         let newConfig = toml.parse(existingConfigStr);
         newConfig.seed_string = seed;
         fs.writeFileSync(setupDefaultsPath, toml.stringify(newConfig));
