+++ src/commands/doge/dummy-signers.ts
@@ -1,0 +1,903 @@
+import * as toml from '@iarna/toml'
+import { confirm, input, select } from '@inquirer/prompts'
+import { Command, Flags } from '@oclif/core'
+import chalk from 'chalk'
+import fs from 'node:fs'
+import path from 'node:path'
+import { execSync } from 'child_process'
+import * as crypto from 'node:crypto'
+import * as os from 'node:os'
+import { fileURLToPath } from 'node:url'
+import { dirname } from 'node:path'
+import type { DogeConfig } from '../../types/doge-config.js'
+import { loadDogeConfig } from '../../utils/doge-config.js'
+import { getSetupDefaultsPath } from '../../config/constants.js'
+
+export class DummySignersManager {
+  private dogeConfig: DogeConfig
+  private configPath: string
+  private log: (message: string) => void
+  private warn: (message: string) => void
+  private error: (message: string) => void
+
+  constructor(
+    dogeConfig: DogeConfig,
+    configPath: string,
+    logger: {
+      log: (message: string) => void
+      warn: (message: string) => void
+      error: (message: string) => void
+    }
+  ) {
+    this.dogeConfig = dogeConfig
+    this.configPath = configPath
+    this.log = logger.log
+    this.warn = logger.warn
+    this.error = logger.error
+  }
+
+  private validators = {
+    required: (value: string) => value.length > 0 ? true : 'This field is required',
+    signerCount: (value: string) => {
+      const num = parseInt(value)
+      return num > 0 && num <= 10 ? true : 'Please enter a number between 1 and 10'
+    }
+  }
+
+  private readTsoHostFromConfig(): string | undefined {
+    try {
+      const configPath = path.resolve(process.cwd(), 'config.toml')
+      if (!fs.existsSync(configPath)) {
+        return undefined
+      }
+      
+      const configContent = fs.readFileSync(configPath, 'utf-8')
+      const config = toml.parse(configContent) as any
+      
+      const tsoHost = config.ingress?.TSO_HOST;
+      
+      if (tsoHost && typeof tsoHost === 'string') {
+        this.log(chalk.blue(`Found TSO_HOST in config.toml: ${tsoHost}`))
+        return tsoHost
+      }
+      
+      return undefined
+    } catch (error) {
+      this.warn(`Failed to read TSO_HOST from config.toml: ${error}`)
+      return undefined
+    }
+  }
+
+  private ensureDirectoryExists(dirPath: string): void {
+    if (!fs.existsSync(dirPath)) {
+      fs.mkdirSync(dirPath, { recursive: true })
+    }
+  }
+
+  private saveConfigToFile(config: any, filePath: string): void {
+    this.ensureDirectoryExists(path.dirname(filePath))
+    fs.writeFileSync(filePath, toml.stringify(config))
+  }
+
+  private async updateSetupDefaultsWithPublicKeys(publicKeys: string[]): Promise<void> {
+    if (publicKeys.length === 0) {
+      this.warn('No public keys were provided');
+      return;
+    }
+
+    const tomlPath = getSetupDefaultsPath();
+    this.ensureDirectoryExists(path.dirname(tomlPath));
+
+    let config: any = {};
+    if (fs.existsSync(tomlPath)) {
+      const tomlContent = fs.readFileSync(tomlPath, 'utf-8');
+      config = toml.parse(tomlContent);
+      this.log(`Loaded existing TOML config from ${tomlPath}`);
+    } else {
+      config = this.getDefaultSetupConfig();
+      this.log('Created new TOML config with defaults');
+    }
+
+    config.correctness_pubkeys = publicKeys;
+
+    const updatedToml = toml.stringify(config);
+    fs.writeFileSync(tomlPath, updatedToml);
+
+    this.log(`✅ Updated ${tomlPath} with ${publicKeys.length} public keys`);
+    this.log(`Public keys: ${publicKeys.join(', ')}`);
+  }
+
+  private getDefaultSetupConfig(): any {
+    return {
+      network: 'testnet',
+      seed_string: '',
+      dogecoin_rpc_url: 'https://testnet.doge.xyz',
+      dogecoin_rpc_user: 'user',
+      dogecoin_rpc_pass: '',
+      dogecoin_blockbook_url: 'https://dogebook-testnet.nownodes.io',
+      dogecoin_blockbook_api_key: '',
+      sequencer_threshold: 1,
+      correctness_threshold: 2,
+      attestation_threshold: 2,
+      recovery_threshold: 2,
+      correctness_key_count: 3,
+      attestation_key_count: 3,
+      recovery_key_count: 3,
+      timelock_seconds: 3600,
+      fee_rate_sat_per_kvb: 2000000,
+      deposit_eth_recipient_address_hex: '0xbb8bc29695232088b1a2dbc117e8c6006478c295',
+      sequencer_target_amount: 42069000,
+      fee_wallet_target_amount: 1000000000,
+      bridge_target_amount: 5000000000,
+      confirmations_required: 1
+    }
+  }
+
+  private async stopAndRemoveContainers(numSigners: number): Promise<void> {
+    this.log(chalk.blue('Cleaning up existing containers...'))
+    for (let i = 0; i < numSigners; i++) {
+      try {
+        execSync(`docker stop dummy-signer-${i}`, { stdio: 'pipe' })
+        this.log(`Stopped dummy-signer-${i}`)
+      } catch {
+        // Container might not exist, ignore
+      }
+      
+      try {
+        execSync(`docker rm dummy-signer-${i}`, { stdio: 'pipe' })
+        this.log(`Removed dummy-signer-${i}`)
+      } catch {
+        // Container might not exist, ignore
+      }
+    }
+  }
+
+  private buildDockerCommand(index: number, config: any, network: string, tsoUrl: string, imageName: string): string {
+    return [
+      'docker run -d',
+      `--name dummy-signer-${index}`,
+      `-p ${config.port}:8080`,
+      `-e DUMMY_SIGNER_WIF="${config.wif}"`,
+      `-e DUMMY_SIGNER_NETWORK="${network}"`,
+      `-e DUMMY_SIGNER_TSO_URL="${tsoUrl}"`,
+      `-e PORT="8080"`,
+      `-e RUST_LOG="info"`,
+      `-e RUST_BACKTRACE="1"`,
+      imageName
+    ].join(' ')
+  }
+
+  async setupDummySigners(): Promise<void> {
+    this.log(chalk.blue('\nSetting up Dummy Signers...'))
+    
+    const deploymentType = await select({
+      message: 'How would you like to run the dummy signers?',
+      choices: [
+        { 
+          name: 'Local (Docker) - For development/testing', 
+          value: 'local',
+          description: 'Run signers locally using Docker with WIF keys'
+        },
+        { 
+          name: 'AWS (App Runner + KMS)', 
+          value: 'aws',
+          description: 'Deploy signers to AWS App Runner with KMS key management'
+        }
+      ],
+      default: this.dogeConfig.deploymentType || 'local'
+    })
+    
+    this.dogeConfig.deploymentType = deploymentType as 'local' | 'aws'
+    
+    if (deploymentType === 'local') {
+      await this.setupLocalSigners()
+    } else {
+      await this.setupAwsSigners()
+    }
+  }
+  
+  private async setupLocalSigners(): Promise<void> {
+    this.log(chalk.blue('\nSetting up Local Dummy Signers...'))
+    
+    try {
+      execSync('docker info', { stdio: 'pipe' })
+    } catch (error) {
+      this.error('Docker is not installed or not running. Please install Docker first.')
+      return
+    }
+    
+    const NUM_SIGNERS = await input({
+      message: 'Number of signers to run locally',
+      default: '3',
+      validate: this.validators.signerCount
+    })
+    
+    const TSO_URL = this.readTsoHostFromConfig()
+    if (!TSO_URL) {
+      this.error('TSO_HOST not found in config.toml. Please add TSO_HOST to your config.toml file.')
+      return
+    }
+    
+    const NETWORK = this.dogeConfig.network
+    
+    const generateWifKeys = await confirm({
+      message: 'Would you like to generate new WIF keys for the signers?',
+      default: true
+    })
+    
+    const signerConfigs = await this.collectSignerConfigs(parseInt(NUM_SIGNERS), generateWifKeys)
+    
+    const imageName = 'dogeos69/dummy-signer:1.0.5-test-local'
+    await this.pullDockerImage(imageName)
+    await this.stopAndRemoveContainers(parseInt(NUM_SIGNERS))
+    await this.startSignerContainers(signerConfigs, NETWORK, TSO_URL, imageName)
+    
+    this.showContainerStatus(signerConfigs)
+    
+    this.saveLocalSignerConfig(NETWORK, signerConfigs)
+    await this.updateSetupDefaultsWithLocalPublicKeys(signerConfigs)
+    
+    this.showSignerUrlsSummary()
+    
+    this.log(chalk.green('\n✅ Local dummy signers setup completed!'))
+  }
+  
+  private async collectSignerConfigs(numSigners: number, generateWifKeys: boolean): Promise<Array<{wif: string, port: number, publicKey?: string}>> {
+    const signerConfigs: Array<{wif: string, port: number, publicKey?: string}> = []
+    
+    for (let i = 0; i < numSigners; i++) {
+      const port = 8080 + i
+      
+      if (generateWifKeys) {
+        this.log(chalk.yellow(`Note: You need to generate WIF key for signer ${i}`))
+      }
+      
+      const wif = await input({
+        message: `Enter WIF private key for signer ${i}`,
+        validate: this.validators.required
+      })
+      
+      signerConfigs.push({ wif, port })
+    }
+    
+    return signerConfigs
+  }
+
+  private async pullDockerImage(imageName: string): Promise<void> {
+    this.log(chalk.blue('Pulling dummy-signer image...'))
+    try {
+      execSync(`docker pull ${imageName}`, { stdio: 'inherit' })
+      this.log(chalk.green('Successfully pulled image'))
+    } catch (error) {
+      this.warn(`Warning: Could not pull image, will try to use local image: ${error}`)
+    }
+  }
+
+  private async startSignerContainers(signerConfigs: any[], network: string, tsoUrl: string, imageName: string): Promise<void> {
+    this.log(chalk.blue(`Starting ${signerConfigs.length} dummy signers...`))
+    
+    for (let i = 0; i < signerConfigs.length; i++) {
+      const config = signerConfigs[i]
+      const dockerCmd = this.buildDockerCommand(i, config, network, tsoUrl, imageName)
+      
+      try {
+        const containerId = execSync(dockerCmd, { encoding: 'utf-8' }).trim()
+        this.log(chalk.green(`✅ Started dummy-signer-${i} on port ${config.port} (container: ${containerId.substring(0, 12)})`))
+      } catch (error) {
+        this.error(`Failed to start dummy-signer-${i}: ${error}`)
+      }
+    }
+  }
+
+  private showContainerStatus(signerConfigs: any[]): void {
+    this.log(chalk.blue('\n📊 Status Summary:'))
+    try {
+      const runningContainers = execSync('docker ps --filter name=dummy-signer --format "table {{.Names}}\\t{{.Status}}\\t{{.Ports}}"', { encoding: 'utf-8' })
+      this.log(runningContainers)
+    } catch (error) {
+      this.warn('Could not get container status')
+    }
+    
+    this.log(chalk.blue('\n🔍 Health check endpoints:'))
+    signerConfigs.forEach((config, i) => {
+      this.log(chalk.blue(`  Signer ${i}: curl http://localhost:${config.port}/health`))
+    })
+    
+    this.log(chalk.blue('\n📋 Useful commands:'))
+    this.log(chalk.blue('  View all containers: docker ps | grep dummy-signer'))
+    this.log(chalk.blue('  View logs: docker logs -f dummy-signer-0'))
+    this.log(chalk.blue('  Stop all: docker stop $(docker ps -q --filter name=dummy-signer)'))
+  }
+
+  private saveLocalSignerConfig(network: string, signerConfigs: any[]): void {
+    if (!this.dogeConfig.localSigners) {
+      this.dogeConfig.localSigners = {}
+    }
+    
+    this.dogeConfig.localSigners = {
+      network,
+      signers: signerConfigs.map((config, i) => ({
+        index: i,
+        port: config.port,
+      }))
+    }
+    
+    // Generate and save local signer URLs
+    this.dogeConfig.signerUrls = signerConfigs.map((config, i) => 
+      `http://localhost:${config.port}`
+    )
+    
+    this.saveConfigToFile(this.dogeConfig, this.configPath)
+    
+    this.log(chalk.green(`Local signer URLs saved: ${this.dogeConfig.signerUrls.join(', ')}`))
+  }
+  
+  private async setupAwsSigners(): Promise<void> {
+    this.log(chalk.blue('\nSetting up AWS Dummy Signers...'))
+    
+    const AWS_REGION = await input({
+      message: 'AWS_REGION',
+      default: this.dogeConfig.awsSigner?.region || 'us-east-1',
+      required: true,
+    })
+    
+    const NETWORK_ALIAS = await input({
+      message: 'NETWORK_ALIAS',
+      default: this.dogeConfig.awsSigner?.networkAlias || 'devnet',
+      required: true,
+    })
+
+    const AWS_ACCOUNT_ID = await input({
+      message: 'AWS_ACCOUNT_ID',
+      required: true,
+      default: this.dogeConfig.awsSigner?.accountId || '',
+    })
+
+    const IMAGE_URI = `${AWS_ACCOUNT_ID}.dkr.ecr.${AWS_REGION}.amazonaws.com/dogeos/dummy-signer:latest`
+
+    const TSO_URL = this.readTsoHostFromConfig()
+    if (!TSO_URL) {
+      this.error('TSO_HOST not found in config.toml. Please add TSO_HOST to your config.toml file.')
+      return
+    }
+
+    const SUFFIXES = await input({
+      message: 'Suffixes for dummy signers (space-separated)',
+      default: this.dogeConfig.awsSigner?.suffixes || '00 01 02',
+      required: false,
+    })
+    
+    await this.saveAwsSignerConfig({
+      region: AWS_REGION,
+      networkAlias: NETWORK_ALIAS,
+      accountId: AWS_ACCOUNT_ID,
+      suffixes: SUFFIXES
+    })
+
+    try {
+      this.prepareDummyImage(AWS_REGION, AWS_ACCOUNT_ID);
+
+      // Get project root directory from current file location
+      const __filename = fileURLToPath(import.meta.url);
+      const __dirname = dirname(__filename);
+      const projectRoot = path.resolve(__dirname, '../../../');
+      const scriptPath = path.join(projectRoot, 'scripts/setup_dummy_signers.sh');
+      
+      // Check if script exists
+      if (!fs.existsSync(scriptPath)) {
+        throw new Error(`Setup script not found at ${scriptPath}`);
+      }
+      
+      this.log(chalk.blue(`Using setup script: ${scriptPath}`));
+      
+      // Execute the script directly with environment variables
+      const cmd = `AWS_REGION=${AWS_REGION} NETWORK_ALIAS=${NETWORK_ALIAS} AWS_ACCOUNT_ID=${AWS_ACCOUNT_ID} IMAGE_URI=${IMAGE_URI} TSO_URL=${TSO_URL} SUFFIXES="${SUFFIXES}" bash ${scriptPath}`;
+      
+      this.log(chalk.blue('Executing AWS setup script...'));
+      execSync(cmd, { stdio: 'inherit' });
+      
+      this.log('Setup dummy signer completed successfully!');
+      
+      await this.updateSetupDefaultsWithKMSPublicKeys(NETWORK_ALIAS, AWS_REGION, SUFFIXES);
+      
+      // Get AWS service URLs
+      await this.getAwsServiceUrls(NETWORK_ALIAS, AWS_REGION, SUFFIXES);
+      
+      this.showSignerUrlsSummary()
+      
+      this.log(chalk.green('\n✅ AWS dummy signers setup completed!'))
+
+    } catch (error) {
+      this.error(`Setup failed: ${error}`);
+    }
+  }
+  
+  private async saveAwsSignerConfig(awsSignerConfig: {
+    region: string
+    networkAlias: string
+    accountId: string
+    suffixes: string
+  }): Promise<void> {
+    try {
+      if (!this.dogeConfig.awsSigner) {
+        this.dogeConfig.awsSigner = {}
+      }
+      
+      this.dogeConfig.awsSigner = {
+        ...this.dogeConfig.awsSigner,
+        ...awsSignerConfig
+      }
+      
+      const configContent = toml.stringify(this.dogeConfig as any)
+      fs.writeFileSync(this.configPath, configContent)
+      
+      this.log(chalk.green(`AWS signer configuration saved to ${this.configPath}`))
+    } catch (error) {
+      this.warn(`Failed to save AWS signer config: ${error}`)
+    }
+  }
+
+  private prepareDummyImage(awsRegion: string, awsAccountId: string): void {
+    const repoName = 'dogeos/dummy-signer';
+    const dockerHubImage = 'dogeos69/dummy-signer:v1.0.5-test';
+    const ecrRegistry = `${awsAccountId}.dkr.ecr.${awsRegion}.amazonaws.com`;
+    const ecrImage = `${ecrRegistry}/${repoName}:latest`;
+    
+    this.log('Checking prerequisites...');
+
+    try {
+      execSync('which aws', { stdio: 'pipe' });
+      execSync('aws sts get-caller-identity', { stdio: 'pipe' });
+    } catch (error) {
+      throw new Error('AWS CLI is not installed or not configured. Please install and configure AWS CLI first.');
+    }
+
+    try {
+      execSync('docker info', { stdio: 'pipe' });
+    } catch (error) {
+      throw new Error('Docker is not installed or not running. Please install and start Docker first.');
+    }
+
+    this.log('Checking ECR repository...');
+    try {
+      execSync(`aws ecr describe-repositories --repository-names ${repoName} --region ${awsRegion}`, { stdio: 'pipe' });
+      this.log(`Repository ${repoName} already exists.`);
+    } catch (error) {
+      this.log(`Repository ${repoName} does not exist. Creating...`);
+      try {
+        execSync(`aws ecr create-repository --repository-name ${repoName} --region ${awsRegion}`, { stdio: 'pipe' });
+        this.log(`Repository ${repoName} created successfully.`);
+      } catch (createError) {
+        throw new Error(`Failed to create ECR repository: ${createError}`);
+      }
+    }
+
+    this.log('Logging in to ECR...');
+    try {
+      execSync(`aws ecr get-login-password --region ${awsRegion} | docker login --username AWS --password-stdin ${ecrRegistry}`, { stdio: 'pipe' });
+      this.log('Successfully logged in to ECR.');
+    } catch (error) {
+      throw new Error(`Failed to log in to ECR: ${error}`);
+    }
+
+    this.log('Checking if image exists in ECR...');
+    let imageExistsInECR = false;
+    try {
+      execSync(`aws ecr describe-images --repository-name ${repoName} --image-ids imageTag=latest --region ${awsRegion}`, { stdio: 'pipe' });
+      imageExistsInECR = true;
+      this.log('Image already exists in ECR.');
+    } catch (error) {
+      this.log('Image does not exist in ECR. Will proceed to push.');
+    }
+
+    if (!imageExistsInECR) {
+      this.log('Checking local Docker images...');
+      let localImageExists = false;
+      try {
+        const result = execSync(`docker images ${dockerHubImage} --format "{{.Repository}}:{{.Tag}}"`, { encoding: 'utf-8' });
+        localImageExists = result.trim() === dockerHubImage;
+      } catch (error) {
+        localImageExists = false;
+      }
+
+      if (!localImageExists) {
+        this.log(`Pulling ${dockerHubImage} from Docker Hub...`);
+        try {
+          execSync(`docker pull ${dockerHubImage}`, { stdio: 'inherit' });
+        } catch (error) {
+          throw new Error(`Failed to pull image from Docker Hub: ${error}`);
+        }
+      } else {
+        this.log('Image already exists locally.');
+      }
+
+      this.log(`Tagging image for ECR...`);
+      try {
+        execSync(`docker tag ${dockerHubImage} ${ecrImage}`, { stdio: 'pipe' });
+      } catch (error) {
+        throw new Error(`Failed to tag image: ${error}`);
+      }
+
+      this.log(`Pushing image to ECR...`);
+      try {
+        execSync(`docker push ${ecrImage}`, { stdio: 'inherit' });
+        this.log('Successfully pushed image to ECR.');
+      } catch (error) {
+        throw new Error(`Failed to push image to ECR: ${error}`);
+      }
+    }
+
+    this.log('Verifying image in ECR...');
+    try {
+      execSync(`aws ecr describe-images --repository-name ${repoName} --image-ids imageTag=latest --region ${awsRegion}`, { stdio: 'pipe' });
+      this.log('✅ Image successfully verified in ECR.');
+    } catch (error) {
+      throw new Error('Failed to verify image in ECR after push.');
+    }
+
+    this.log('Dummy signer image preparation completed successfully!');
+  }
+
+  private async updateSetupDefaultsWithKMSPublicKeys(networkAlias: string, awsRegion: string, suffixesStr: string): Promise<void> {
+    try {
+      this.log('Fetching KMS public keys...');
+      
+      const suffixes = suffixesStr.split(' ').filter(s => s.trim());
+      const publicKeys: string[] = [];
+      
+      for (const suffix of suffixes) {
+        const aliasName = `alias/${networkAlias}-dummy-signer-${suffix}-key`;
+        
+        try {
+          const keyIdOutput = execSync(
+            `aws kms describe-key --key-id "${aliasName}" --region ${awsRegion} --query KeyMetadata.KeyId --output text`,
+            { encoding: 'utf-8' }
+          ).trim();
+          
+          const publicKeyOutput = execSync(
+            `aws kms get-public-key --key-id "${keyIdOutput}" --region ${awsRegion} --query PublicKey --output text`,
+            { encoding: 'utf-8' }
+          ).trim();
+          
+          const publicKeyHex = this.convertKMSPublicKeyToHex(publicKeyOutput);
+          publicKeys.push(publicKeyHex);
+          
+          this.log(`✅ Got public key for ${aliasName}: ${publicKeyHex}`);
+        } catch (error) {
+          this.warn(`Failed to get public key for ${aliasName}: ${error}`);
+        }
+      }
+      
+      await this.updateSetupDefaultsWithPublicKeys(publicKeys);
+      
+    } catch (error) {
+      this.error(`Failed to update setup defaults: ${error}`);
+    }
+  }
+  
+  private convertKMSPublicKeyToHex(base64PublicKey: string): string {
+    const derBuffer = Buffer.from(base64PublicKey, 'base64');
+    
+    this.log(`DER buffer (first 100 bytes): ${derBuffer.subarray(0, 100).toString('hex')}`);
+    
+    let bitStringStart = -1;
+    for (let i = 0; i < derBuffer.length - 2; i++) {
+      if (derBuffer[i] === 0x03) {
+        const length = derBuffer[i + 1];
+        if (length === 0x42 || length === 0x43) {
+          bitStringStart = i;
+          break;
+        }
+      }
+    }
+    
+    if (bitStringStart === -1) {
+      for (let i = 0; i < derBuffer.length - 65; i++) {
+        if (derBuffer[i] === 0x04 && i + 65 <= derBuffer.length) {
+          const possibleKey = Buffer.from(derBuffer.subarray(i, i + 65));
+          const x = possibleKey.subarray(1, 33);
+          const y = possibleKey.subarray(33, 65);
+          
+          const xSum = x.reduce((a, b) => a + b, 0);
+          const ySum = y.reduce((a, b) => a + b, 0);
+          
+          if (xSum > 0 && xSum < 255 * 32 && ySum > 0 && ySum < 255 * 32) {
+            const prefix = (y[31] & 1) === 0 ? 0x02 : 0x03;
+            const compressedKey = Buffer.concat([Buffer.from([prefix]), Buffer.from(x)]);
+            return compressedKey.toString('hex');
+          }
+        }
+      }
+      
+      throw new Error('Could not find valid public key in DER format');
+    }
+    
+    const publicKeyStart = bitStringStart + 3;
+    
+    if (derBuffer[publicKeyStart] !== 0x04) {
+      throw new Error('Expected uncompressed public key (0x04 prefix)');
+    }
+    
+    const uncompressedKey = Buffer.from(derBuffer.subarray(publicKeyStart, publicKeyStart + 65));
+    
+    const x = uncompressedKey.subarray(1, 33);
+    const y = uncompressedKey.subarray(33, 65);
+    
+    const prefix = (y[31] & 1) === 0 ? 0x02 : 0x03;
+    
+    const compressedKey = Buffer.concat([Buffer.from([prefix]), Buffer.from(x)]);
+    
+    return compressedKey.toString('hex');
+  }
+
+  private async updateSetupDefaultsWithLocalPublicKeys(signerConfigs: Array<{wif: string, port: number, publicKey?: string}>): Promise<void> {
+    try {
+      this.log('Fetching public keys from WIF...');
+      
+      const publicKeys: string[] = [];
+      
+      for (const config of signerConfigs) {
+        const publicKey = this.extractPublicKeyFromWIF(config.wif);
+        publicKeys.push(publicKey);
+      }
+      
+      await this.updateSetupDefaultsWithPublicKeys(publicKeys);
+      
+    } catch (error) {
+      this.error(`Failed to update setup defaults: ${error}`);
+    }
+  }
+  
+  private extractPublicKeyFromWIF(wif: string): string {
+    try {
+      const base58Alphabet = '123456789ABCDEFGHJKLMNPQRSTUVWXYZabcdefghijkmnopqrstuvwxyz';
+      
+      const decoded = this.base58Decode(wif, base58Alphabet);
+      
+      if (decoded.length < 37) {
+        throw new Error('Invalid WIF length');
+      }
+      
+      const payload = Buffer.from(decoded.subarray(0, -4));
+      const checksum = Buffer.from(decoded.subarray(-4));
+      const hash1 = crypto.createHash('sha256').update(payload).digest();
+      const hash2 = crypto.createHash('sha256').update(hash1).digest();
+      const expectedChecksum = Buffer.from(hash2.subarray(0, 4));
+      
+      if (!checksum.equals(expectedChecksum)) {
+        throw new Error('Invalid WIF checksum');
+      }
+      
+      const privateKey = Buffer.from(payload.subarray(1, 33));
+      
+      const isCompressed = payload.length === 34;
+      
+      const publicKey = this.generatePublicKeyFromPrivate(privateKey, isCompressed);
+      
+      return publicKey;
+    } catch (error) {
+      throw new Error(`Failed to extract public key from WIF: ${error}`);
+    }
+  }
+  
+  private base58Decode(input: string, alphabet: string): Buffer {
+    const base = alphabet.length;
+    let decoded = BigInt(0);
+    let multi = BigInt(1);
+    
+    for (let i = input.length - 1; i >= 0; i--) {
+      const char = input[i];
+      const index = alphabet.indexOf(char);
+      if (index === -1) {
+        throw new Error(`Invalid character in base58 string: ${char}`);
+      }
+      decoded += BigInt(index) * multi;
+      multi *= BigInt(base);
+    }
+    
+    const hex = decoded.toString(16);
+    const paddedHex = hex.length % 2 === 0 ? hex : '0' + hex;
+    let buffer = Buffer.from(paddedHex, 'hex');
+    
+    for (let i = 0; i < input.length && input[i] === alphabet[0]; i++) {
+      buffer = Buffer.concat([Buffer.from([0]), buffer]);
+    }
+    
+    return buffer;
+  }
+  
+  private generatePublicKeyFromPrivate(privateKey: Buffer, compressed: boolean = true): string {
+    try {
+      const ecdh = crypto.createECDH('secp256k1');
+      ecdh.setPrivateKey(privateKey);
+      
+      const publicKey = ecdh.getPublicKey();
+      
+      if (compressed) {
+        const x = Buffer.from(publicKey.subarray(1, 33));
+        const y = Buffer.from(publicKey.subarray(33, 65));
+        const prefix = (y[31] & 1) === 0 ? 0x02 : 0x03;
+        const compressedKey = Buffer.concat([Buffer.from([prefix]), x]);
+        return compressedKey.toString('hex');
+      } else {
+        return publicKey.toString('hex');
+      }
+    } catch (error) {
+      throw new Error(`Failed to generate public key: ${error}`);
+    }
+  }
+
+  private async getAwsServiceUrls(networkAlias: string, awsRegion: string, suffixesStr: string): Promise<void> {
+    try {
+      this.log('Fetching AWS App Runner service URLs...');
+      
+      const suffixes = suffixesStr.split(' ').filter(s => s.trim());
+      const urls: string[] = [];
+      
+      for (const suffix of suffixes) {
+        const serviceName = `${networkAlias}-dummy-signer-${suffix}`;
+        
+        try {
+          // Get the service ARN first
+          const serviceListOutput = execSync(
+            `aws apprunner list-services --region ${awsRegion} --query "ServiceSummaryList[?ServiceName=='${serviceName}'].ServiceArn" --output text`,
+            { encoding: 'utf-8' }
+          ).trim();
+          
+          if (serviceListOutput) {
+            // Get the service URL using the ARN
+            const serviceUrlOutput = execSync(
+              `aws apprunner describe-service --service-arn "${serviceListOutput}" --region ${awsRegion} --query "Service.ServiceUrl" --output text`,
+              { encoding: 'utf-8' }
+            ).trim();
+            
+            if (serviceUrlOutput && serviceUrlOutput !== 'None') {
+              const fullUrl = serviceUrlOutput.startsWith('https://') ? serviceUrlOutput : `https://${serviceUrlOutput}`;
+              urls.push(fullUrl);
+              this.log(`✅ Got service URL for ${serviceName}: ${fullUrl}`);
+            } else {
+              this.warn(`Service URL not found for ${serviceName}`);
+            }
+          } else {
+            this.warn(`Service ${serviceName} not found`);
+          }
+        } catch (error) {
+          this.warn(`Failed to get service URL for ${serviceName}: ${error}`);
+        }
+      }
+      
+      if (urls.length > 0) {
+        this.dogeConfig.signerUrls = urls;
+        this.saveConfigToFile(this.dogeConfig, this.configPath);
+        this.log(chalk.green(`AWS service URLs saved: ${urls.join(', ')}`));
+      } else {
+        this.warn('No AWS service URLs found. Services may still be deploying.');
+      }
+      
+    } catch (error) {
+      this.error(`Failed to get AWS service URLs: ${error}`);
+    }
+  }
+
+  private showSignerUrlsSummary(): void {
+    this.log(chalk.blue('\n📊 Signer URLs Summary:'))
+    if (this.dogeConfig.signerUrls && this.dogeConfig.signerUrls.length > 0) {
+      this.dogeConfig.signerUrls.forEach((url, index) => {
+        this.log(chalk.blue(`  Signer ${index}: ${url}`))
+      })
+    } else {
+      this.log(chalk.yellow('  No signer URLs found'))
+    }
+  }
+}
+
+// Command class for oclif CLI
+export class DummySignersCommand extends Command {
+  static description = 'Set up dummy signers (local Docker or AWS with KMS keys)'
+
+  static examples = [
+    '$ scrollsdk doge:dummy-signers',
+    '$ scrollsdk doge:dummy-signers --config .data/doge-config-testnet.toml',
+    '$ scrollsdk doge:dummy-signers --local-only',
+    '$ scrollsdk doge:dummy-signers --aws-only',
+  ]
+
+  static flags = {
+    config: Flags.string({
+      char: 'c',
+      description: 'Path to config file (e.g., .data/doge-config-mainnet.toml or .data/doge-config-testnet.toml)',
+    }),
+    'local-only': Flags.boolean({
+      char: 'l',
+      description: 'Set up local Docker signers only',
+      default: false,
+    }),
+    'aws-only': Flags.boolean({
+      char: 'a',
+      description: 'Set up AWS KMS signers only',
+      default: false,
+    }),
+  }
+
+  private dogeConfig: DogeConfig = {} as DogeConfig
+  private configPath: string = ''
+
+  async run(): Promise<void> {
+    const { flags } = await this.parse(DummySignersCommand)
+
+    // Resolve config path
+    let resolvedPath = ""
+    if (!flags.config) {
+      // Auto-detect config file
+      if (!fs.existsSync('.data')) {
+        this.error('No .data directory found. Please run `scrollsdk doge:config` first to create a configuration.')
+        return
+      }
+      
+      const files = fs.readdirSync('.data')
+      const configFiles = files.filter(file => file.endsWith('.toml'))
+      
+      if (configFiles.length === 0) {
+        this.error('No config files found in .data directory. Please run `scrollsdk doge:config` first.')
+        return
+      }
+      
+      if (configFiles.length === 1) {
+        resolvedPath = path.resolve('.data', configFiles[0])
+        this.log(chalk.blue(`Using config file: ${resolvedPath}`))
+      } else {
+        this.error(`Multiple config files found: ${configFiles.join(', ')}. Please specify one with --config flag.`)
+        return
+      }
+    } else {
+      resolvedPath = path.resolve(flags.config)
+      if (!fs.existsSync(resolvedPath)) {
+        this.error(`Config file ${resolvedPath} does not exist`)
+        return
+      }
+    }
+
+    // Load config
+    this.configPath = resolvedPath
+    try {
+      this.dogeConfig = await loadDogeConfig(resolvedPath)
+      this.log(chalk.blue(`Loaded configuration for ${this.dogeConfig.network} network`))
+    } catch (error) {
+      this.error(`Failed to load config: ${error}`)
+      return
+    }
+
+    // Validate flags
+    if (flags['local-only'] && flags['aws-only']) {
+      this.error('Cannot use both --local-only and --aws-only flags together')
+      return
+    }
+
+    // Override deployment type if flags are specified
+    if (flags['local-only']) {
+      this.dogeConfig.deploymentType = 'local'
+    } else if (flags['aws-only']) {
+      this.dogeConfig.deploymentType = 'aws'
+    }
+
+    // Set up dummy signers
+    const dummySignersManager = new DummySignersManager(
+      this.dogeConfig,
+      this.configPath,
+      {
+        log: this.log.bind(this),
+        warn: this.warn.bind(this),
+        error: this.error.bind(this)
+      }
+    )
+    
+    try {
+      await dummySignersManager.setupDummySigners()
+    } catch (error) {
+      this.error(`Dummy signers setup failed: ${error}`)
+    }
+  }
+}
+
+// Export the command as default for oclif
+export default DummySignersCommand
