+++ tsconfig.json
@@ -1,13 +1,16 @@
 {
   "compilerOptions": {
     "declaration": true,
-    "module": "Node16",
+    "module": "ESNext",
     "outDir": "dist",
     "rootDir": "src",
     "strict": true,
     "target": "es2022",
-    "moduleResolution": "node16",
-    "lib": ["es2022", "dom"]
+    "moduleResolution": "bundler",
+    "lib": ["es2022", "dom"],
+    "skipLibCheck": true,
+    "allowSyntheticDefaultImports": true,
+    "esModuleInterop": true
   },
   "include": ["./src/**/*"],
   "ts-node": {
